NDExCard extracted
